chore: move the test case for sql pair deletion
